+++ CHANGELOG.md
@@ -4,6 +4,7 @@ History of **rest2go** library with changes description.
 
 ## Next release
 
+- `FEATURE` Log HTTP request & response middleware
 - `FEATURE` REST API errors handling
 - `FEATURE` Loaded once settings as YAML file
 - `FEATURE` GoLang library project initialization